Moved GetSearchableMailboxes class and function to API module
